fixup! Add stats-tracking querier for cardinality metrics
